Add chalky output

+++ cli.js
@@ -2,6 +2,8 @@
 'use strict';
 const meow = require('meow');
 const fs = require('fs');
+const chalk = require('chalk');
+const {log} = console;
 
 const cli = meow(`
 
@@ -25,14 +27,27 @@ const cli = meow(`
 
 if (cli.flags.override) {
 	if (!fs.existsSync('./src')) {
+		log(chalk.blue('Creating src directory'));
 		fs.mkdirSync('./src');
 	}
 
 	if (cli.flags.override === 'html') {
-		fs.copyFileSync('./templates/index.html', './src/index.html');
+		if (fs.existsSync('./src/index.html')) {
+			log(chalk.red('index.html already exists!'));
+		} else {
+			fs.copyFileSync('./templates/index.html', './src/index.html');
+			log(chalk.green('index.html created successfully.'));
+		}
 	} else if (cli.flags.override === 'vue') {
-		fs.copyFileSync('./templates/App.vue', './src/App.vue');
+		if (fs.existsSync('./src/App.vue')) {
+			log(chalk.red('App.vue already exists!'));
+		} else {
+			fs.copyFileSync('./templates/App.vue', './src/App.vue');
+			log(chalk.green('App.vue created successfully.'));
+		}
+	} else {
+		console.log(chalk.yellow('Please provide "html" or "vue" to the --override option.'));
 	}
 } else {
-  cli.showHelp()
+	cli.showHelp();
 }
